Refactor AudioService and rework Player classes
